Refactor XliffConverter
To fix formatting issue, check for a < or &, and if the source or target contains one, create a CDATA section, otherwise create a text node.
Remove requirement for needing extra information about what we're converting from and just use some defaults instead. This information is not saved in PHP or YAML, so makes converting between xliff and those formats tricky for getting the correct information. Because we no longer need those, remove those from the ConverterInterface.

Fixes #3